Refactor: 리액트 쿼리 설정 중복 요소 통일
