+++ src/app/ReactQueryProvider.tsx
@@ -1,11 +1,17 @@
 'use client';
 
+import queryClientOption from '@/query/config';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { PropsWithChildren, useState } from 'react';
 
 const ReactQueryProvider = ({ children }: PropsWithChildren) => {
-  const [queryClient] = useState(() => new QueryClient());
+  const [queryClient] = useState(
+    () =>
+      new QueryClient({
+        defaultOptions: queryClientOption.defaultOptions,
+      }),
+  );
 
   return (
     <QueryClientProvider client={queryClient}>
