--- src/lib/api/testApi/index.ts
@@ -1,7 +1,0 @@
-import { apiPrivate } from '../config';
-const onTest = async () => {
-  const { data } = await apiPrivate.post('/test');
-  return data;
-};
-
-export default onTest;
