+++ src/query/config/index.ts
@@ -1,0 +1,11 @@
+const queryClientOption = {
+  defaultOptions: {
+    queries: {
+      staleTime: 1000 * 60,
+      cacheTime: 1000 * 60 * 5,
+      retry: false,
+    },
+  },
+};
+
+export default queryClientOption;
